Separate the server and client component for SSR of a blog post - integrate the checkBlogs function, get rid of json dependency
